Move test specific code to __tests__ directory, use jsdom environment everywhere
