+++ src/index.ts
@@ -22,7 +22,6 @@ export * from "./Goban/CanvasRenderer";
 export * from "./Goban/SVGRenderer";
 export * from "./Goban/themes";
 export * from "./Goban/Goban";
-export * from "./Goban/TestGoban";
 
 export * as protocol from "engine/protocol";
 export { placeRenderedImageStone, preRenderImageStone } from "./Goban/themes/image_stones";
